# specs/openapi/dist/api-bundle.yaml
openapi: 3.1.0
info:
  title: Virtual Companion API
  version: 0.1.0
  description: Contract-first HTTP API (ADR-0002). This document is the single source of truth; Java interfaces and TypeScript
    clients are generated from it and must not be hand-edited or used to back-drive the contract. The ErrorEnvelope shape
    is uniform across every error response; NOT_FOUND_OR_FORBIDDEN hides resource existence so a missing resource and a forbidden
    one are indistinguishable to the caller.
security:
- bearerAuth: []
paths:
...
  /api/v1/exports:
    post:
      operationId: createExport
      summary: Enqueue an asynchronous data export (FR-DATA-002)
      description: Creates an export request and enqueues a DATA_EXPORT work item (vc.create_export_request); the worker builds
        the JSON document asynchronously and the status endpoint reports PENDING until READY. Only one in-flight export per
        account is allowed. The response issues the one-time download token EXACTLY ONCE (only its sha256 digest is stored,
        V76); save downloadUrl immediately — status polls never carry it again. Requires a server-verified authentication
        context.
      responses:
        '200':
          description: The created export request (PENDING)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportRequest'
        '400':
          description: An export for this account is already in flight
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: RATE_LIMITED — shared per-owner export admission rejected the request
          headers:
            Retry-After:
              description: Seconds until the caller should retry
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exports/{exportId}:
    get:
      operationId: getExportStatus
      summary: Read one owned export request status (FR-DATA-002)
      description: Status of the caller's export request via vc.get_export_request. The response never carries the download
        token or URL (the token was issued once in the create response, V76); a foreign or absent id maps to NOT_FOUND_OR_FORBIDDEN
        (existence is never disclosed). Requires a server-verified authentication context.
      parameters:
      - $ref: '#/components/parameters/ExportId'
      responses:
        '200':
          description: The export request status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportRequest'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Export request absent or not owned; existence is not disclosed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/exports/{exportId}/download:
    get:
      operationId: downloadExport
      summary: One-time authenticated download of a READY export (FR-DATA-002)
      description: Returns the export document exactly once (the token is consumed on success, vc.consume_export digests the
        presented token before comparison). The token was issued once in the create response; expired, consumed, foreign or
        absent exports map to NOT_FOUND_OR_FORBIDDEN (existence is never disclosed). Requires a server-verified authentication
        context.
      parameters:
      - $ref: '#/components/parameters/ExportId'
      - $ref: '#/components/parameters/ExportToken'
      responses:
        '200':
          description: The export document (one-time)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportDownload'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not READY, expired, consumed, absent or not owned; existence is not disclosed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorEnvelope:
      type: object
      description: Unified error shape. NOT_FOUND_OR_FORBIDDEN never discloses whether a resource exists; every error response
        uses this envelope.
      required:
      - code
      - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Stable, non-sensitive human-readable explanation
        details:
          type: object
          additionalProperties: true
          description: Optional structured context (never sensitive data)
    ErrorCode:
      type: string
      description: Stable error code catalog (specs/catalog/error-codes.yaml)
      enum:
      - AUTHENTICATION_REQUIRED
      - ACCESS_DENIED
      - NOT_FOUND_OR_FORBIDDEN
      - AGE_VERIFICATION_REQUIRED
      - MODEL_NO_ELIGIBLE_DEPLOYMENT
      - MODEL_ATTEMPT_TIMEOUT
      - MODEL_ALL_UNAVAILABLE
      - GENERATION_CANCELLED
      - GENERATION_OUTPUT_BLOCKED
      - STREAM_GAP_EXPIRED
      - STREAM_RESET_REQUIRED
      - MEMORY_CONFIRMATION_REQUIRED
      - MEMORY_FORBIDDEN_CONTENT
      - CATALOG_DRIFT_DETECTED
      - BETA_OPERATIONS_NOT_READY
      - INVALID_REQUEST
      - AUTH_RATE_LIMITED
...
    ExportRequest:
      type: object
      description: One asynchronous data-export request (FR-DATA-002).
      required:
      - exportId
      - status
      - requestedAt
      properties:
        exportId:
          type: string
          description: Stable owner-scoped export request identifier
        status:
          type: string
          description: PENDING, READY, FAILED or EXPIRED
          enum:
          - PENDING
          - READY
          - FAILED
          - EXPIRED
        requestedAt:
          type: string
          format: date-time
          description: Request timestamp (RFC 3339)
        completedAt:
          type: string
          format: date-time
          description: Completion timestamp, present once READY
        expiresAt:
          type: string
          format: date-time
          description: Download expiry, present once READY
        errorMessage:
          type: string
          description: Stable failure reason, present only on FAILED
        downloadToken:
          type: string
          description: One-time download secret, present ONLY in the create response (V76); only its sha256 digest is stored
            server-side.
        downloadUrl:
          type: string
          description: Short-lived one-time download link, present ONLY in the create response; status polls never repeat
            it.
    ExportDownload:
      type: object
      description: The exported user-data document (FR-DATA-002, one-time).
      required:
      - exportId
      - generatedAt
      - expiresAt
      - aiContentNotice
      - conversations
      - memories
      - reminders
      - consents
      properties:
        exportId:
          type: string
          description: The export request identifier
        generatedAt:
          type: string
          format: date-time
          description: Document generation timestamp (RFC 3339)
        expiresAt:
          type: string
          format: date-time
          description: Document expiry timestamp (RFC 3339)
        aiContentNotice:
          type: string
          description: Plain-language notice that assistant messages are AI-generated
        conversations:
          type: array
          items:
            $ref: '#/components/schemas/ExportConversation'
        memories:
          type: array
          items:
            $ref: '#/components/schemas/ExportMemory'
        reminders:
          type: array
          items:
            $ref: '#/components/schemas/ExportReminder'
        consents:
          type: array
          items:
            $ref: '#/components/schemas/ExportConsent'
    ExportConversation:
      type: object
      description: One conversation with its full message history (FR-DATA-002).
      required:
      - conversationId
      - relationshipId
      - incognito
      - messages
      properties:
        conversationId:
          type: string
          description: Owner-scoped conversation identifier
        relationshipId:
          type: string
          description: Owning relationship identifier
        title:
          type: string
          description: User-renamed title, absent until renamed
        incognito:
          type: boolean
          description: Frozen creation-time incognito flag
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ExportMessage'
    ExportMessage:
      type: object
      description: One message with its AI-generated marker (FR-DATA-002).
      required:
      - messageId
      - role
      - content
      - aiGenerated
      properties:
        messageId:
          type: string
          description: Owner-scoped message identifier
        role:
          type: string
          description: user or assistant
        content:
          type: string
          description: The message text
        aiGenerated:
          type: boolean
          description: true for assistant messages (AI 生成内容标识)
    ExportMemory:
      type: object
      description: One memory record (FR-DATA-002).
      required:
      - memoryId
      - scope
      - summary
      - status
      - createdAt
      properties:
        memoryId:
          type: string
          description: Owner-scoped memory identifier
        relationshipId:
          type: string
          description: Owning relationship identifier, present for RELATIONSHIP scope
        scope:
          type: string
          description: memory-scopes catalog code
        summary:
          type: string
          description: Memory content summary
        status:
          type: string
          description: memory-candidate-statuses catalog code
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (RFC 3339)
        deletedAt:
          type: string
          format: date-time
          description: Deletion timestamp, present only when deleted
    ExportReminder:
      type: object
      description: One structured reminder (FR-DATA-002).
      required:
      - reminderId
      - relationshipId
      - text
      - remindAt
      - recurrence
      - status
      properties:
        reminderId:
          type: string
          description: Owner-scoped reminder identifier
        relationshipId:
          type: string
          description: Owning relationship identifier
        text:
          type: string
          description: The reminder text
        remindAt:
          type: string
          format: date-time
          description: UTC instant the reminder is due
        recurrence:
          type: string
          description: NONE, DAILY or WEEKLY
        status:
          type: string
          description: ACTIVE or DISMISSED
    ExportConsent:
      type: object
      description: One effective consent record (FR-DATA-002).
      required:
      - consentId
      - consentType
      - version
      - granted
      - grantedAt
      properties:
        consentId:
          type: string
          description: Owner-scoped consent record identifier
        consentType:
          type: string
          description: The FR-AUTH-003 consent type
        version:
          type: string
          description: The policy document version
        granted:
          type: boolean
          description: Effective granted state
        grantedAt:
          type: string
          format: date-time
          description: Record timestamp (RFC 3339)
        revokedAt:
          type: string
          format: date-time
          description: Revocation timestamp, present only on a revoke record
...
  parameters:
...
    ExportId:
      name: exportId
      in: path
      required: true
      schema:
        type: string
      description: Owner-scoped export request identifier
    ExportToken:
      name: token
      in: query
      required: true
      schema:
        type: string
      description: The one-time download token handed out inside the authenticated status response